+++ src/brski/pledge/pledge_request.cpp
@@ -194,6 +194,10 @@ std::string create_cert_string(const char *cert) {
 
 int generate_sign_cert(struct BinaryArray *scert_cert,
                        struct BinaryArray *scert_key) {
+  uint8_t rand[8];
+  char rands[17];
+  struct BinaryArray buf = {.array = rand, .length = 8};
+
   struct crypto_cert_meta sign_cert_meta = {
       .serial_number = 12345,
       .not_before = 0,
@@ -202,6 +206,13 @@ int generate_sign_cert(struct BinaryArray *scert_cert,
       .issuer = NULL,
       .subject = NULL,
       .basic_constraints = (char *)"CA:false"};
+
+  if (crypto_getrand(&buf) < 0) {
+    log_error("crypto_getrand fail");
+    return -1;
+  }
+
+  printf_hex(rands, 16, rand, 8, 1);
 
   if ((sign_cert_meta.issuer = init_keyvalue_list()) == NULL) {
     log_error("init_keyvalue_list fail");
@@ -221,7 +232,13 @@ int generate_sign_cert(struct BinaryArray *scert_cert,
   }
 
   if (push_keyvalue_list(sign_cert_meta.subject, (char *)"CN",
-                         (char *)"pledge-domain-cert") < 0) {
+                         (char *)"ldevid-cert") < 0) {
+    log_error("push_keyvalue_list fail");
+    goto generate_sign_cert_err;
+  }
+
+  if (push_keyvalue_list(sign_cert_meta.subject, (char *)"serialNumber",
+                         rands) < 0) {
     log_error("push_keyvalue_list fail");
     goto generate_sign_cert_err;
   }